Refactor NotificationToggle into a settings component
