+++ client/src/actionCreators/userSettings.js
@@ -1,4 +1,4 @@
-import { TOGGLE_NOTIFICATION } from '../../../common/actionTypes/notification';
+import { TOGGLE_NOTIFICATION } from '../../../common/actionTypes/userSettings';
 
 // eslint-disable-next-line import/prefer-default-export
 export const toggleNotification = () => ({
