+++ client/src/components/Admin/Home.js
@@ -3,7 +3,7 @@ import DocumentTitle from 'react-document-title';
 import { IssueContainer } from './Issue';
 import Alternatives from './Alternatives';
 import { ConcludedIssueListContainer } from '../ConcludedIssueList';
-import NotificationToggle from '../NotificationToggle';
+import UserSettings from '../UserSettings';
 import IssueStatus from '../IssueStatus';
 import css from '../../css/Home.css';
 
@@ -18,7 +18,7 @@ const AdminHome = () => (
         </div>
         <IssueStatus />
       </div>
-      <NotificationToggle />
+      <UserSettings />
       <ConcludedIssueListContainer />
     </div>
   </DocumentTitle>
