--- client/src/selectors/notification.js
@@ -1,3 +1,0 @@
-export const notificationIsEnabled = state => state.notification.enabled;
-
-export default notificationIsEnabled;
