+++ client/src/selectors/userSettings.js
@@ -1,0 +1,3 @@
+export const notificationIsEnabled = state => state.userSettings.notifications;
+
+export default notificationIsEnabled;
