+++ common/actionTypes/userSettings.js
